+ organise virtual ram related  + cleanup from buggy flash based imp
