+++ CMakeLists.txt
@@ -166,7 +166,6 @@ else ()
 
             # supports only for pico with 4 or more MB flashes
             PICO_FLASH_SIZE_BYTES=4194304
-            #PSEUDO_RAM_BASE=0x10300000
             SD_CARD_SWAP=1
     )
