+++ src/main.cpp
@@ -39,103 +39,6 @@ bool runing = true;
 
 #if PICO_ON_DEVICE
 
-#if SD_CARD_SWAP
-static const char* path = "\\XT\\pagefile.sys";
-static FIL file;
-
-bool init_vram() {
-    logMsg((char *)"Create <SD-card>\\XT\\pagefile.sys");
-    FRESULT result = f_open(&file, path, FA_READ | FA_WRITE | FA_CREATE_ALWAYS);
-    if (result == FR_OK) {
-        result = f_lseek(&file, TOTAL_VIRTUAL_MEMORY_KBS * 1024);
-        if (result != FR_OK) {
-            logMsg((char *)"Unable to init <SD-card>\\XT\\pagefile.sys");
-            return false;
-        }
-    } else {
-        logMsg((char *)"Unable to create <SD-card>\\XT\\pagefile.sys");
-        return false;
-    }
-    f_close(&file);
-    result = f_open(&file, path, FA_READ | FA_WRITE);
-    if (result != FR_OK) {
-        logMsg((char *)"Unable to open <SD-card>\\XT\\pagefile.sys");
-    }
-    logMsg((char *)"pagefile.sys is initialized");
-    return true;
-}
-
-FRESULT vram_seek(FIL* fp, uint32_t file_offset) {
-    FRESULT result = f_lseek(&file, file_offset);
-    if (result != FR_OK) {
-        result = f_open(&file, path, FA_READ | FA_WRITE);
-        if (result != FR_OK) {
-            char tmp[40];
-            sprintf(tmp, "Unable to open pagefile.sys: %s (%d)", FRESULT_str(result), result);
-            logMsg(tmp);
-            return result;
-        }
-        char tmp[40];
-        sprintf(tmp, "Failed to f_lseek: %s (%d)", FRESULT_str(result), result);
-        logMsg(tmp);
-    }
-    return result;
-}
-
-void read_vram_block(char* dst, uint32_t file_offset, uint32_t sz) {
-    gpio_put(PICO_DEFAULT_LED_PIN, true);
-    char tmp[40];
-    //sprintf(tmp, "Read  pagefile 0x%X<-0x%X", dst, file_offset);
-    //logMsg(tmp);
-    FRESULT result = vram_seek(&file, file_offset);
-    if (result != FR_OK) {
-        return;
-    }
-    UINT br;
-    result = f_read(&file, dst, sz, &br);
-    if (result != FR_OK) {
-        sprintf(tmp, "Failed to f_read: %s (%d)", FRESULT_str(result), result);
-        logMsg(tmp);
-    }
-    gpio_put(PICO_DEFAULT_LED_PIN, false);
-}
-
-void flush_vram_block(const char* src, uint32_t file_offset, uint32_t sz) {
-    gpio_put(PICO_DEFAULT_LED_PIN, true);
-    char tmp[40];
-    //sprintf(tmp, "Flush pagefile 0x%X->0x%X", src, file_offset);
-    //logMsg(tmp);
-    FRESULT result = vram_seek(&file, file_offset);
-    if (result != FR_OK) {
-        return;
-    }
-    UINT bw;
-    result = f_write(&file, src, sz, &bw);
-    if (result != FR_OK) {
-        sprintf(tmp, "Failed to f_write: %s (%d)", FRESULT_str(result), result);
-        logMsg(tmp);
-    }
-    gpio_put(PICO_DEFAULT_LED_PIN, false);
-}
-#endif
-
-#if PSEUDO_RAM_BASE
-#include <hardware/flash.h>
-// TODO: own C file
-void flash_range_program3(uint32_t addr, const u_int8_t * buff, size_t sz) {
-    gpio_put(PICO_DEFAULT_LED_PIN, true);
-    char tmp[40]; // sprintf(tmp, "Flash erase 0x%X (%d)", addr, sz); logMsg(tmp);
-    uint32_t interrupts = save_and_disable_interrupts();
-    // multicore_lockout_start_blocking();
-    flash_range_erase(addr - XIP_BASE, sz);
-    sprintf(tmp, "Flash write 0x%X<-0x%X", addr, buff); logMsg(tmp);
-    flash_range_program(addr - XIP_BASE, buff, sz);
-    // multicore_lockout_end_blocking();
-    restore_interrupts(interrupts);
-    gpio_put(PICO_DEFAULT_LED_PIN, false);
-}
-#endif
-
 struct semaphore vga_start_semaphore;
 /* Renderer loop on Pico's second core */
 void __time_critical_func(render_core)() {
